feat: added `EQUALS` to MultiVersion enum   affected classes due to import changes:    - PacketWrapper, ServerVersion, WrapperPlayWorldBorderWarningDelay
